IOS-1828 Initial changes to support mercuryo
